move eval_ppl to a parent script
